implementata la generazione casuale dei musei, delle visite e la composizione casuali di pacchetti per il test PopolaDB. Scritti i test per i servizi di recupero delle visite e fixato il problema con la data delle visite. implementati i test fin'ora ignorati. il file unporotte.csv contiene rotte che hanno almeno dei voli che partono ed arrivano nelle stesse città.

+++ TravelDreamX/TravelDreamX-ejb/src/main/java/it/polimi/traveldream/service/ParametriRicercaPB.java
@@ -37,7 +37,7 @@ public class ParametriRicercaPB {
 
     public void setGiornoVisita(Date giornoVisita) {
         long time = giornoVisita.getTime();
-        time = ((long)time / (24*60*60*1000))*(24*60*60*1000);
+        time = ((long)time / (long)(24*60*60*1000))*(24*60*60*1000);
         this.giornoVisita = new Date(time);
     }
